requirements.txt
asgiref==3.3.1
Django==3.1.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==14.2.0
django-filter==2.4.0
gunicorn==20.0.4
pytz==2021.1
sqlparse==0.4.1
